requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
docker-pycreds==0.4.0
filelock==3.18.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.0
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
tinydb==4.8.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ucimlrepo==0.0.7
urllib3==2.4.0
wandb==0.19.11
matplotlib==3.10.3
seaborn==0.13.2
typer==0.15.1
sqlalchemy==2.0.41
torchmetrics==1.7.1
wget==3.2